fix: no pyinstaller flare-floss (#3)
The Floss binary uses pyinstaller which seems to struggle with cleaning
up its binaries.

Now that we are on a newer version of Python it seems that we can
directly install the Floss binary and that works fine.

Subprocess management/timeout remains the same as a new Python
interpreter is spawned to get this working.